Separating in files the communication classes
