+++ hardwarelibrary/communication/communicationport.py
@@ -3,6 +3,7 @@ import re
 import time
 import random
 from threading import Thread, RLock
+from commands import *
 
 class CommunicationReadTimeout(serial.SerialException):
     pass
@@ -12,124 +13,6 @@ class CommunicationReadNoMatch(Exception):
 
 class CommunicationReadAlternateMatch(Exception):
     pass
-
-class Command:
-    def __init__(self, name:str, endPoints = (None, None)):
-        self.name = name
-        self.reply = None
-        self.matchGroups = None
-        self.endPoints = endPoints
-
-        self.isSent = False
-        self.isSentSuccessfully = False
-        self.exceptions = []
-
-        self.isReplyReceived = False
-        self.isReplyReceivedSuccessfully = False
-
-    def matchAsFloat(self, index=0):
-        if self.matchGroups is not None:
-            return float(self.matchGroups[index])
-        return None
-    
-    @property
-    def hasError(self):
-        return len(self.exceptions) != 0
-
-    def send(self, port) -> bool:
-        raise NotImplementedError()
-
-class TextCommand(Command):
-    def __init__(self, name, text, replyPattern = None, 
-                                   alternatePattern = None, 
-                                   multiReplyCount = 1,
-                                   finalReplyPattern = None, 
-                                   endPoints = (None, None)):
-        Command.__init__(self, name, endPoints=endPoints)
-        self.text : str = text
-        self.replyPattern: str = replyPattern
-        self.alternatePattern: str = alternatePattern
-        self.multiReplyCount = multiReplyCount
-        self.finalReplyPattern = finalReplyPattern
-
-    def send(self, port, params=None) -> bool:
-        try:
-            self.isSent = True
-            if params is not None:
-                textCommand = self.text.format(params)
-            else:
-                textCommand = self.text
-
-            port.writeString(string=textCommand, endPoint=self.endPoints[0])
-
-            if self.multiReplyCount > 1:
-                self.reply = []
-                self.matchGroups = []
-
-                for i in range(self.multiReplyCount):
-                    reply, matchGroups = port.readMatchingGroups(
-                                   replyPattern=self.replyPattern,
-                                   alternatePattern=self.alternatePattern,
-                                   endPoint=self.endPoints[1])
-                    self.reply.append(reply)
-                    self.matchGroups.append(matchGroups)
-
-            elif self.finalReplyPattern is not None:
-                self.reply = []
-                self.matchGroups = []
-
-                while True:
-                    reply, matchGroups = port.readMatchingGroups(
-                                   replyPattern=self.replyPattern,
-                                   alternatePattern=self.alternatePattern,
-                                   endPoint=self.endPoints[1])
-                    self.reply.append(reply)
-                    self.matchGroups.append(matchGroups)
-
-                    if re.search(self.finalReplyPattern, reply) is not None:
-                        break
-            else:
-                if self.replyPattern is not None:
-                    self.reply, self.matchGroups = port.readMatchingGroups(
-                               replyPattern=self.replyPattern,
-                               alternatePattern=self.alternatePattern,
-                               endPoint=self.endPoints[1])
-                else:
-                    pass
-            
-            self.isSentSuccessfully = True
-        except Exception as err:
-            self.exceptions.append(err)
-            self.isSentSuccessfully = False
-            return True
-
-        return False
-
-
-class DataCommand(Command):
-    def __init__(self, name, data, replyHexRegex = None, replyDataLength = 0, unpackingMask = None, endPoints = None):
-        Command.__init__(self, name, endPoints=endPoints)
-        self.data : bytearray = data
-        self.replyHexRegex: str = replyHexRegex
-        self.replyDataLength: int = replyDataLength
-        self.unpackingMask:str = unpackingMask
-
-    def send(self, port) -> bool:
-        try:
-            self.isSent = True
-            nBytes = port.writeData(data=self.data, endPoint=self.endPoints[0])
-            if self.replyDataLength > 0:
-                self.reply = port.readData(length=self.replyDataLength)
-            elif self.replyHexRegex is not None:
-                raise NotImplementedError()
-                # self.reply = port.readData(length=self.replyDataLength)
-            self.isSentSuccessfully = True
-        except Exception as err:
-            self.exceptions.append(err)
-            self.isSentSuccessfully = False
-            return True
-
-        return False
 
 class CommunicationPort:
     """CommunicationPort class with basic application-level protocol 
@@ -145,6 +28,10 @@ class CommunicationPort:
     @property
     def isOpen(self):
         raise NotImplementedError()
+
+    @property
+    def isNotOpen(self):
+        return not self.isOpen
 
     def open(self):
         raise NotImplementedError()
@@ -230,50 +117,3 @@ class CommunicationPort:
         else:
             raise CommunicationReadNoMatch("Unable to match pattern:'{0}' in reply:'{1}'".format(replyPattern, reply))
 
-
-class DebugEchoCommunicationPort(CommunicationPort):
-    def __init__(self, delay=0):
-        self.buffers = [bytearray(),bytearray()]
-        self.delay = delay
-        self._isOpen = False
-        super(DebugEchoCommunicationPort, self).__init__()
-
-    @property
-    def isOpen(self):
-        return self._isOpen    
-
-    def open(self):
-        if self._isOpen:
-            raise Exception()
-
-        self._isOpen = True
-        return
-
-    def close(self):
-        self._isOpen = False
-        return
-
-    def bytesAvailable(self, endPoint=0):
-        return len(self.buffers[endPoint])
-
-    def flush(self):
-        self.buffers = [bytearray(),bytearray()]
-
-    def readData(self, length, endPoint=0):
-        with self.portLock:
-            time.sleep(self.delay*random.random())
-            data = bytearray()
-            for i in range(0, length):
-                if len(self.buffers[endPoint]) > 0:
-                    byte = self.buffers[endPoint].pop(0)
-                    data.append(byte)
-                else:
-                    raise CommunicationReadTimeout("Unable to read data")
-
-        return data
-
-    def writeData(self, data, endPoint=0):
-        with self.portLock:
-            self.buffers[endPoint].extend(data)
-
-        return len(data)
